fix database url issue

+++ utils/schema.js
@@ -5,7 +5,7 @@ export const Budgets = pgTable('budgets', {
     user_id: varchar('user_id').notNull(),
     name: varchar('name').notNull(),
     icon: varchar('icon'),
-    amount: numeric('amount').notNull(),
+    amount: integer('amount').notNull(),
     createdBy: varchar('createdBy').notNull(),
 })
 
@@ -13,7 +13,7 @@ export const Expenses = pgTable('expenses', {
     id: serial('id').primaryKey(),
     name: varchar('name').notNull(),
     user_id: varchar('user_id').notNull(),
-    amount: numeric('amount').notNull().default(0),
+    amount: integer('amount').notNull().default(0),
     budgetId: integer('budgetId').references(() => Budgets.id),
     createdAt: varchar('createdAt').notNull(),
 })
